code refactors based off main PR suggestions
