+++ src/components/TaskListPage/airpax/utils/index.js
@@ -1,0 +1,9 @@
+import BaggageUtil from './baggageUtil';
+import DateTimeUtil from './datetimeUtil';
+import IndicatorsUtil from './indicatorsUtil';
+import BookingUtil from './bookingUtil';
+import DocumentUtil from './documentUtil';
+import PersonUtil from './personUtil';
+import MovementUtil from './movementUtil';
+
+export { BaggageUtil, DateTimeUtil, IndicatorsUtil, BookingUtil, DocumentUtil, PersonUtil, MovementUtil };
